refactor: share more input value/event helpers
Extract more duplicated input logic into utils/input:

- buildInputResponseEvent: the answered/selected event plus
  answerAncestor context, previously copied across booleanInput,
  choiceInput, mathInput, textInput, codeEditor, and the matrix/fraction
  shared helper.
- returnImmediateValueStateVariableDefinition: the immediateValue state
  variable (mirrors value, holds in-progress edits) shared by textInput
  and codeEditor.
- inputUpdateImmediateValue / inputUpdateValue: the commit-on-blur/enter
  actions shared by textInput and codeEditor.

No behavior change; all input vitest suites pass.
