+++ packages/doenetml-worker-javascript/src/components/ChoiceInput.js
@@ -3,7 +3,10 @@ import me from "math-expressions";
 import { enumerateCombinations, enumeratePermutations } from "@doenet/utils";
 import { setUpVariantSeedAndRng } from "../utils/variants";
 import { returnListItemChildStateVariableDefinitions } from "../utils/listItemChild";
-import { defineSubmitAnswerExternalAction } from "../utils/input";
+import {
+    buildInputResponseEvent,
+    defineSubmitAnswerExternalAction,
+} from "../utils/input";
 
 export default class Choiceinput extends Input {
     constructor(args) {
@@ -1582,26 +1585,14 @@ export default class Choiceinput extends Input {
             ];
 
             let choiceTexts = await this.stateValues.choiceTexts;
-            let event = {
+            let event = await buildInputResponseEvent({
+                component: this,
                 verb: "selected",
-                object: {
-                    componentIdx: this.componentIdx,
-                    componentType: this.componentType,
-                },
-                result: {
-                    response: effectiveSelectedIndices,
-                    responseText: effectiveSelectedIndices.map(
-                        (i) => choiceTexts[i - 1],
-                    ),
-                },
-            };
-
-            let answerAncestor = await this.stateValues.answerAncestor;
-            if (answerAncestor) {
-                event.context = {
-                    answerAncestor: answerAncestor.componentIdx,
-                };
-            }
+                response: effectiveSelectedIndices,
+                responseText: effectiveSelectedIndices.map(
+                    (i) => choiceTexts[i - 1],
+                ),
+            });
 
             await this.coreFunctions.performUpdate({
                 updateInstructions,
