+++ packages/doenetml-worker-javascript/src/components/TextInput.js
@@ -8,6 +8,9 @@ import { returnTextPieceStateVariableDefinitions } from "../utils/text";
 import Input from "./abstract/Input";
 import {
     defineSubmitAnswerExternalAction,
+    inputUpdateImmediateValue,
+    inputUpdateValue,
+    returnImmediateValueStateVariableDefinition,
     returnInputValueChangedStateVariableDefinitions,
 } from "../utils/input";
 
@@ -16,8 +19,8 @@ export default class Textinput extends Input {
         super(args);
 
         Object.assign(this.actions, {
-            updateImmediateValue: this.updateImmediateValue.bind(this),
-            updateValue: this.updateValue.bind(this),
+            updateImmediateValue: inputUpdateImmediateValue.bind(this),
+            updateValue: inputUpdateValue.bind(this),
             moveInput: this.moveInput.bind(this),
         });
 
@@ -301,91 +304,12 @@ export default class Textinput extends Input {
             },
         };
 
-        stateVariableDefinitions.immediateValue = {
-            description:
-                "The text value reflecting the user's in-progress edits.",
-            public: true,
-            shadowingInstructions: {
+        stateVariableDefinitions.immediateValue =
+            returnImmediateValueStateVariableDefinition({
+                description:
+                    "The text value reflecting the user's in-progress edits.",
                 createComponentOfType: "text",
-            },
-            forRenderer: true,
-            hasEssential: true,
-            shadowVariable: true,
-            returnDependencies: () => ({
-                value: {
-                    dependencyType: "stateVariable",
-                    variableName: "value",
-                },
-                immediateValueChanged: {
-                    dependencyType: "stateVariable",
-                    variableName: "immediateValueChanged",
-                    onlyToSetInInverseDefinition: true,
-                },
-            }),
-            definition: function ({
-                dependencyValues,
-                changes,
-                justUpdatedForNewComponent,
-                usedDefault,
-            }) {
-                // console.log(`definition of immediateValue`)
-                // console.log(dependencyValues)
-                // console.log(changes);
-
-                if (
-                    changes.value &&
-                    !justUpdatedForNewComponent &&
-                    !usedDefault.value
-                ) {
-                    // only update to value when it changes
-                    // (otherwise, let its essential value change)
-                    return {
-                        setValue: { immediateValue: dependencyValues.value },
-                        setEssentialValue: {
-                            immediateValue: dependencyValues.value,
-                        },
-                    };
-                } else {
-                    return {
-                        useEssentialOrDefaultValue: {
-                            immediateValue: {
-                                defaultValue: dependencyValues.value,
-                            },
-                        },
-                    };
-                }
-            },
-            inverseDefinition: function ({
-                desiredStateVariableValues,
-                initialChange,
-                shadowedVariable,
-            }) {
-                // value is essential; give it the desired value
-                let instructions = [
-                    {
-                        setEssentialValue: "immediateValue",
-                        value: desiredStateVariableValues.immediateValue,
-                    },
-                    {
-                        setDependency: "immediateValueChanged",
-                        desiredValue: true,
-                    },
-                ];
-
-                // if from outside sources, also set value
-                if (!(initialChange || shadowedVariable)) {
-                    instructions.push({
-                        setDependency: "value",
-                        desiredValue: desiredStateVariableValues.immediateValue,
-                    });
-                }
-
-                return {
-                    success: true,
-                    instructions,
-                };
-            },
-        };
+            }).immediateValue;
 
         stateVariableDefinitions.text = {
             description: "The current text as a plain text string.",
@@ -413,106 +337,6 @@ export default class Textinput extends Input {
         Object.assign(stateVariableDefinitions, pieceDefs);
 
         return stateVariableDefinitions;
-    }
-
-    async updateImmediateValue({
-        text,
-        actionId,
-        sourceInformation = {},
-        skipRendererUpdate = false,
-    }) {
-        if (!(await this.stateValues.disabled)) {
-            return await this.coreFunctions.performUpdate({
-                updateInstructions: [
-                    {
-                        updateType: "updateValue",
-                        componentIdx: this.componentIdx,
-                        stateVariable: "immediateValue",
-                        value: text,
-                    },
-                    {
-                        updateType: "setComponentNeedingUpdateValue",
-                        componentIdx: this.componentIdx,
-                    },
-                ],
-                transient: true,
-                actionId,
-                sourceInformation,
-                skipRendererUpdate,
-            });
-        }
-    }
-
-    async updateValue({
-        actionId,
-        sourceInformation = {},
-        skipRendererUpdate = false,
-    }) {
-        if (!(await this.stateValues.disabled)) {
-            let immediateValue = await this.stateValues.immediateValue;
-
-            if ((await this.stateValues.value) !== immediateValue) {
-                let updateInstructions = [
-                    {
-                        updateType: "updateValue",
-                        componentIdx: this.componentIdx,
-                        stateVariable: "value",
-                        value: immediateValue,
-                    },
-                    // in case value ended up being a different value than requested
-                    // we set immediate value to whatever was the result
-                    // (hence the need to execute update first)
-                    // Also, this makes sure immediateValue is saved to the database,
-                    // since in updateImmediateValue, immediateValue is not saved to database
-                    {
-                        updateType: "executeUpdate",
-                    },
-                    {
-                        updateType: "updateValue",
-                        componentIdx: this.componentIdx,
-                        stateVariable: "immediateValue",
-                        valueOfStateVariable: "value",
-                    },
-                    {
-                        updateType: "unsetComponentNeedingUpdateValue",
-                    },
-                ];
-
-                let event = {
-                    verb: "answered",
-                    object: {
-                        componentIdx: this.componentIdx,
-                        componentType: this.componentType,
-                    },
-                    result: {
-                        response: immediateValue,
-                        responseText: immediateValue,
-                    },
-                };
-
-                let answerAncestor = await this.stateValues.answerAncestor;
-                if (answerAncestor) {
-                    event.context = {
-                        answerAncestor: answerAncestor.componentIdx,
-                    };
-                }
-
-                await this.coreFunctions.performUpdate({
-                    updateInstructions,
-                    actionId,
-                    sourceInformation,
-                    skipRendererUpdate: true,
-                    event,
-                });
-
-                return await this.coreFunctions.triggerChainedActions({
-                    componentIdx: this.componentIdx,
-                    actionId,
-                    sourceInformation,
-                    skipRendererUpdate,
-                });
-            }
-        }
     }
 
     async moveInput({
